Introduce: ks-nav tool set for Linux kernel analysis
This commit introduces the ks-nav tool set, designed to support ELISA
activities related to Linux kernel analysis. The tool set includes two
complementary tools that work together to provide a comprehensive solution.

The first tool is responsible for collecting information from various
sources and storing them into a database. The second tool uses the database
to generate diagrams that aid in analyzing the Linux kernel.

The tool set introduces several new functionalities, such as labels to arcs
in nav subsystems mode, targeted view mode, references to the code for indirect
calls, and a switch to indicate "stop at" or "stop before". These
functionalities improve the analysis process and provide more flexibility to
the user.

This initial commit include a fair test coverage.

The tool set is specifically designed for Linux kernel analysis and is intended
to provide a more comprehensive and efficient solution for ELISA activities.
This commit marks the introduction of the tool set, and we look forward to
continued development and improvements in the future.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>
Signed-off-by: Maurizio Papini <maurizio.papini@gmail.com>
Signed-off-by: Lev Veyde <lveyde@gmail.com>
Signed-off-by: Leonardo Rossetti <lrossett@redhat.com>
Signed-off-by: Veronika Fuxova <vfuxova@redhat.com>
Signed-off-by: Rob Linden <rokkbert@gmail.com>